Build dependencies. Hate life.
Figuring out how to build libzmq was surprisingly difficult. The
documentation says it's as simple as `autogen.sh`, `./configure`,
`make`, `make install`, but I determined that to be a lie. It failed to
build `curve_keygen` because it wasn't linking against the C++ std
library. So I had to add

    --enable-shared --enable-static LDFLAGS="-lstdc++ -lm" --prefix=$(INSTALL_DIR)

to the configure script incantation.

Building zmqpp is much simpler, but note that it requires libzmq, so you
have to modify CXXFLAGS and LDFLAGS, in addition to changing the
installation prefix.